sign_task_response refactored
