+++ crates/operator/Cargo.toml
@@ -20,6 +20,7 @@ eyre.workspace = true
 futures-util.workspace = true
 
 # eigen-rs
+eigen-aggregator.workspace = true
 eigen-client-avsregistry.workspace = true
 eigen-config.workspace = true
 eigen-crypto-bls.workspace = true
